karigai_backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Redis for caching
redis==5.2.1
hiredis==3.3.0

# AI/ML Services
openai==1.59.5
google-cloud-vision==3.8.1
google-cloud-translate==3.18.0

# Audio processing
openai-whisper==20240930
pydub==0.25.1
librosa==0.10.2.post1
scipy==1.15.1
numpy==2.2.1

# Document generation
reportlab==4.2.5
weasyprint==63.1
jinja2==3.1.5

# Image processing
pillow==11.1.0
opencv-python==4.10.0.84

# HTTP client
httpx==0.28.1
aiohttp==3.11.11

# ElevenLabs TTS
elevenlabs==1.15.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.1
celery==5.4.0
pytest==8.3.4
pytest-asyncio==0.25.2

# Logging and monitoring
structlog==24.4.0
sentry-sdk[fastapi]==2.20.0

# File handling
aiofiles==24.1.0
python-magic==0.4.27

# WhatsApp integration
twilio==9.4.0

# Development
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.14.1
hypothesis==6.125.0